Do not panic on incorrect input
We have a public method that panics if the input is invalid, this is
unnecessary we should return an error.

Add a new error variant `ArrayLengthsNotEqual`, use it and the existing
`NoTransactions` for the error paths instead of panicing.

While we are changing the function signature, take ownership of the
`header` parameter because we use it to build the returned data
structure - taking ownership makes more sense than borrowing and
copying.