Standardize derives on error types
Currently we use a variety of derives when creating error types. Since
error types are part of the public API we want to get to a state where
we can commit to not changing them.

Select a set of derived traits and standardize all our error types by
deriving this set. Of note, `Copy` is included in the chosen set of
derives which means we cannot have variants with an inner `io::Error`,
instead use `io::ErrorKind`.

Standardize derived traits on error types to be:

`#[derive(Clone, Debug, PartialEq, Eq)]`

Includes conversion of `io::ErrorKind` to `io::Error` in order to get at
a `Display` implementation (`io::ErrorKind` does not implement `Display`
until Rust 1.60).