Use dtonlnay instead of actions-rs
Currently we use the `actions-rs` GitHub action to run our tests. It
seems the project is now unmaintained [0].

Well known Rust developer dtonlnay maintains a GitHub action that can be
used instead.

Replace all uses of `actions-rs/toolchain` with
`dtonlnay/rust-toolchain`. Note that with the new action there is no way
to configure the toolchain, instead a different `uses` statement is
required - this means we have to split our jobs up by toolchain. This is
arguably cleaner anyways.

Note that with this patch applied the "no-std" tests are now _not_ run
for MSRV since we explicitly support "no-std" only for the 1.47 and
above toolchains - strange that this was working?

[0] https://github.com/actions-rs/toolchain/issues/216