Split `Hash` trait
The `Hash` trait requires for the hash type to support hashing arbitrary
data. This is not great, since hash newtypes usually hash specific data.
Hashing arbitrary data into newtypes is usually a bug or
misunderstanding of the API.

This change separates methods performing hashing of arbitrary data to a
separate `RawHash` trait and provides a few helpers to make usage
easier:

* adds a *private* `engine()` method to the hash newtype
* adds `finalize()` method to `HashEngine`
* adds `hashes::hash` function

Additionally, because of visibility issues, this moves many hash
newtypes from the dedicated `hash_types` module to the modules that
produce said hashes.

Closes #1481