Patch the bitcoin_hashes dependency
Currently we have a hole in our dependency graph that is making it
impossible to patch `bitcoin_hashes` with breaking changes.

  bitcoin -> secp -> hashes
          -> hashes

We tried to fix this by using a `path` dependency for `hashes` but this
does not work.

Use a git/rev dependency for `hashes` that makes `bitcoin` depend on a
recent commit on the master branch.

With this applied we will be able to make breaking changes to `hashes`
and still build `bitcoin`.