Update secp256k1 dependency
We just merged into `secp256k1` a change to make secp depend on a
git/ref version of `hashes`. This was aimed at resolving the dependency
hole issue we currently have, we can now use the new version of secp.
This completes the dependency hole fix.

FTR with this merged the merge workflow for making breaking changes to
`hashes` can be:

1. Patch `hashes`
2. Wait for merge of 1, then patch secp (incl. update of git/rev dep)
3. Wait for merge of 2, then patch bitcoin (incl. update of git/rev dep)

One dev workflow can be:
0. Patch hashes and secp dependencies to use path
1. Make the changes everywhere
2. Make three separate commits that will make up the steps 1-3 above