Depend on the new hex crate
We recently created a `hex` crate that copied code from
`bitcoin_internals` and `hashes::hex`.

Depend on the new `hex` crate and remove the `internals::hex` and
`hashes::hex` modules.

Note, adds reexport of `hex` from `hashes` so we can use it in macros
and from `bitcoin` to improve ergonomics. This technically means a bunch
of the imports could have been left as `crate::hashes::hex` but elect to
change them to the more terse and obvious `crate::hex`.

Includes update to `Cargo-minimal.lock` and `Cargo-latest.lock`.