Rename TapSighashHash to TapSighash
The `TapSighash` is the taproot sighash, it is a little misnamed as it
is. Change the name to `TapSighash`.

Leave the definition in the `taproot` module. This is a little
incongruent to its non-taproot counterpart `hash_types::Sighash`. I have
no better idea for where to put these currently.

Fix: #1550