Improve to/from_consensus
We have many types that convert to/from consensus encode integers. We
would like to provide an API that is:

- Intuitive (i.e., uniform method names)
- Ergonomic for hardcore Bitcoin/Rust devs
- Safe and clear without forcing policy on downstream devs

Attempt to achieve the above by adding additional methods to various
types. Also add examples to help document the methods and as rustdoc
tests (found we were lacking in this department using `cargo mutants`).

With this applied all relevant types have, (or should have assuming I
haven't missed any):

- from_consensus
- From<type> or TryFrom<type>
- to_consensus_<type>
- to_<type>

Some have `u8` and `u32` for <type>