Remove no-std feature
A few observations:

- `rust-bitcoin` requires an allocator, it is not an optional feature.

- We have a requirement that users pick either "std" or "no-std" (or
both) to build however there is no obvious reason why we should not be
able to build with `cargo check --no-default-features` other than the
requirement to configure the allocator.

- `core2` re-exports stdlib if the "std" feature is enabled.

With the above observations we can add "core2" as a non-optional
dependency (with "alloc" feature). We can then enable "core2/std" if our
"std" feature is enabled. This allows us to remove the "no-std" feature
all together.