script: Improve inlining
In general `#[inline]` should only be used on functions that, when
compiled to assembler, are a few instructions long.

To make explicit my incomplete knowledge:

I believe that calling functions across crate boundaries is slower than
calling functions within a crate because the compiler does not have as
much chance to optimize the code - that is the extent of my
understanding on that but it made me think that there is no point
inlining a function that calls to a non-inlined function from our crate.

Go through the `script` module and think a bit harder about how many
assembly instructions a function will compile down to. Remove and add
`inline` as appropriate.