Add check_pub_error macro
Public errors are part of the public API, as such we need to be careful
not to change them. One thing that is hard to notice is adding/removing
`Send`/`Sync`.

Add a macro that asserts that an error type implements `Send` and
`Sync`. While we are at it, assert `Debug` is implemented as well
because we would like all errors to implement `Debug`.