Split the ecdsa signature types
Currently we have signatures and sighash types that are specific to the
ECDSA signing algorithm. Observe however that the segwit v0 sighash type
commits to different data to the legacy sighash type - the signatures
and sighash types are not interchangeable. We can use the type system to
enforce this difference by having multiple types.

As an initial attempt, duplicate the `ecdsa::Signature` and
`EcdasSighashType` types (and associated code). Leave the ecdsa stuff in
place because it is currently used by the PSBT module and teasing this
apart is non-trivial.

Fix the rest of the codebase (excl. `psbt`) to not use the old ECDSA
types any more.