Introduce primitives crate
We would like to provide a crate that is the `rust-bitcoin` primitives
types, this crate should allow users (specifically `rust-miniscript`) to
depend solely on `primitives` and not on `rust-bitcoin`.

The big picture is that we will then be able to invert the dependency
between `miniscript` and `bitcoin` - making `rust-bitcoin` the
one-stop-shop for all things Bitcoin (in Rust).

The other primary problem we are trying to solve is the upgrade friction
the ecosystem feels every time we release `rust-bitcoin`. Currently many
downstream crates have to upgrade in sync. With `primitives` we hope to
provide a crate that:

- Changes less often
- Stabilizes more quickly

Hence reducing some of the friction - time will tell.

Notes on `crypto` / `secp256k1`
-------------------------------

`secp256k1` is required for key types. Currently we cannot provide key
types in pure rust because serializing and deserializing them requires
libsecp, although it is technically possible that we could do all that
in pure Rust, as of today we don't and it is a non-trivial amount of
work to do so.

As such, the `primitives` crate has a dependency on `secp256k1`, we
feature gate it behind a "crypto" feature. It should be noted however
that it is unlikely that there will be many users whom are able use
`primitives` without enabling the "crypto" feature. This means that
the `primitives` crate is not a silver bullet in respect to the upgrade
friction mentioned above because `secp256k1` is someways off stabalizing
- it does still mean that being smaller it is less likely to change.
Whether we will ever manage to do a `rust-bitcoin` release without a
`primitives` release remains to be seen but there is reasonable chance.