Pull witness_version out of `WitnessProgram`
The version byte appears _before_ the witness program in a segwit
scriptPubkey however our `WitnessProgram` struct includes it. Either the
struct is incorrectly named of the version byte should not be part of
it. Observe also that we typically use the version byte at the same time
as the witness program and also that it is required to do the program
length checks.

Pull the version byte out of `WitnessProgram` and inline it into the
`Payload` segwit variant after first renaming the variant from
`Payload::WitnessProgram` to `Payload::Segwit`.

This change makes the code more semantically correct. Note that it does
introduce a non-explicit assumption that the same version be passed to
`WitnessProgram::new` as is used in the `Payload` - is this acceptable
engineering?