Add a script to check the public API
We would like to check for API changes during development and in CI so
that such changes can be discussed separately from the code. We have API
listings already.

Add tooling and documentation for creating API listings for the public
crates within the workspace (i.e., not `internals`).

Add API text files from an initial run of the script.