policy: Add refactor carve out
I have managed to burn out or bore our reviewers/maintainers. Getting
two acks is becoming increasingly difficult. I've pestered everyone to
the limit that I feel socially comfortable doing so am requesting a
carve out to the 2-ACK before merge rule.

The primary justification is that I feel we should have a bit more of
BDFL and a bit less total consensus if we are to push forwards.