Improve the address error handling
The `address` currently has a single large error enum that is returned
by all functions irrespective of whether the function can fail with all
variants or not.

Improve the address error handling by doing:

- Create a general purpose error type `address::Error` that has a
variant for each of the other error variants in the module. This type is
not currently returned by any of our functions but is provided for users
of the library who wish to call various `address` functions and do not
care to match on the specific error type.

- Create specific error types for functions so that each function
returns an error that is either a struct or an enum in which all
variants are used.