Use vec instead of ECDSA sig in psbt input
Currently we use `ecdsa::Signature` in `Input`. The PSBT is a
serialization format so we can use a vector of bytes with no loss of
functionality.

Also, we would like to split the `ecdsa` types into legacy and segwit
specific types and using the `ecdsa::Signature` type in PSBT make this
very difficult.

Serialize the ECDSA signature into a vector of bytes when storing it in
the PSBT `Input` type.

This changes the `serde` format of the serialized PSTB which means we
update the serde regression test in this patch also.

Note that the PSBT serialization format as specified by BIP-174 is not
changed by this patch. Evident by passing tests and no code changes to
`tests/psbt.rs`.