Remove use of classify in Instructions::next
We are abusing the `classify` function to check if an opcode is a data
push, we can do this instead by simply adding an `push_bytes` function
to `Opcode` and then directly comparing against the opcode for the
remaining `OP_PUSHDATA*` opcodes.

Done in preparation for removing the `Class` and `Ordinary` types.