opcodes: Remove Class, Ordinary, and classify
We no longer use the `Class` or `Ordinary` types or the
`Opcode::classify` function. This code can all be removed.

This change removes public types and functions.

Also this change removes the classify logic and mention of
legacy/taproot context. Note that taproot context was never used in our
code.